Testing mailer

+++ app/controllers/favorites_controller.rb
@@ -2,6 +2,7 @@ class FavoritesController < ApplicationController
   def create
     post = Post.find(params[:post_id])
     favorite = current_user.favorites.build(post: post)
+    authorize favorite
 
     if favorite.save
       flash[:notice] = "Successfully Favorited!"
@@ -15,6 +16,7 @@ class FavoritesController < ApplicationController
   def destroy
     post = Post.find(params[:post_id])
     favorite = current_user.favorites.find(params[:id])
+    authorize favorite
 
     if favorite.destroy
       flash[:notice] = "Post unfavorited successfully."
